requirements.txt
factory-boy==2.10.0
Faker==0.8.11
ipaddress==1.0.19
python-dateutil==2.6.1
six==1.11.0
text-unidecode==1.2
Django==1.11.10
factory-boy==2.10.0
Faker==0.8.11
ipaddress==1.0.19
MySQL-python==1.2.5
python-dateutil==2.6.1
pytz==2018.3
six==1.11.0
text-unidecode==1.2
